+++ packages/server/bff-runtime/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/bff-runtime
 
+## 2.4.0
+
 ## 2.3.0
 
 ## 2.2.0